+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Current
 New Features
 ^^^^^^^^^^^^
  - Add dataset for patch sampling from (large) nD images without loading entire images into memory (:gh: `806` by `Vicky De Ridder`_)
+ - Add learned regularizers and nonmonotonic accelerated proximal gradient algorithm (:gh:`829` by `Johannes Hertrich`_)
 
 Changed
 ^^^^^^^